adding more specific tests for output
DWARF is also funny in that it (given a header) will generate a function
(subprogram) twice and one has the correct (named) formal params and the
other does not, so I have to do some matching. The same is true for arrays -
the actual name is with the member and parsing the array outside of that
context does not make sense.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>