refactor: preparing to add compatibility (image and application) to compspec!
We want this library to extend beyond basic asp graphs (diffs primarily) and
into our current work for compatibility! I am really excited about this
idea because I can add a plugin approach to allow for someone to define
a custom compspec plugin (in Python) as an extractor, and then it can
still easily push to a registry (as an artifact) with oras. E.g., we can
have compspec-ior for IOR, and it can be installed as a separate python
module plugin here to extract the specification. I will next likely add
the plugin framework, jsonschema to validate a schema, and then similar
functionality to derive an extractor using a plugin (so likely an interface,
etc). I wanted to do more this weekend but ran out of time working on
other things, but will hopefully do more here soon :)

Signed-off-by: vsoch <vsoch@users.noreply.github.com>